src/requirements.txt
boto3
awscli-local
PyYAML